consumer: V2 consumers implementation
v2/consumers defines consumption in terms of ShardSpecs, which model the
specific configuration to be consumed. It defines the Consumer and Store
interfaces which applications must implement, and the shard processing
lifecycle which powers consumers.

It provides a simple JSON Store and a RocksDB store, though applications
are not limited to these implementations.

It also implements a gRPC ShardServer API for interacting with consumer shards
via the gazctl tool, and a `Resolver` to aide applications in resolving ShardIDs
to local instances or proxy-able Routes.